% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_gene_drug_response.R
\name{analyze_gene_drug_response}
\alias{analyze_gene_drug_response}
\title{Analyze Association between Gene (Signature) and Drug Response}
\usage{
analyze_gene_drug_response(gene_list, combine = FALSE)
}
\arguments{
\item{gene_list}{a gene symbol list.}

\item{combine}{if \code{TRUE}, combine the expression of gene list as
a gene signature.}
}
\value{
a \code{data.frame}
}
\description{
Analyze Association between Gene (Signature) and Drug Response
}
\examples{
\dontrun{
analyze_gene_drug_response("TP53")
analyze_gene_drug_response(c("TP53", "KRAS"))
analyze_gene_drug_response(c("TP53", "KRAS"), combine = TRUE)
}
}
