% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_gene_drug_response2.R
\name{analyze_gene_drug_response_asso2}
\alias{analyze_gene_drug_response_asso2}
\title{Another Function to Analyze Association between Gene (Signature)
and Drug Response with CCLE Data}
\usage{
analyze_gene_drug_response_asso2(
  gene_list,
  response_type = c("IC50", "EC50", "Amax", "ActArea"),
  tissue = "ALL",
  manova = FALSE,
  combine = FALSE
)
}
\arguments{
\item{gene_list}{a gene symbol list.}

\item{response_type}{one measure type for drug response.}

\item{tissue}{a tissue name. Check examples.}

\item{combine}{if \code{TRUE}, combine the expression of gene list as
a gene signature.}
}
\value{
a \code{data.frame}
\itemize{
\item If \code{combine} is \code{TRUE}, genes are combined as \code{signature}.
\item \code{mean.diff} and \code{median.diff} indicate mean and median of
normalized expression difference between High IC50 cells and Low IC50 cells.
The cutoff between High and Low are median IC50.
}
}
\description{
Analyze gene-drug association using ANOVA or MANOVA.
}
\examples{
\dontrun{
analyze_gene_drug_response_asso2("TP53")
analyze_gene_drug_response_asso2(c("TP53", "KRAS"))
analyze_gene_drug_response_asso2(c("TP53", "KRAS"), combine = TRUE)
analyze_gene_drug_response_asso2("TP53", response_type = "EC50")
analyze_gene_drug_response_asso2("TP53", response_type = "Amax")
analyze_gene_drug_response_asso2("TP53", response_type = "IC50", manova = TRUE)

# Visualization
vis_gene_drug_response_asso2("TP53")
}
}
\references{
Garnett, Mathew J., et al. "Systematic identification of genomic markers of drug sensitivity in cancer cells." Nature 483.7391 (2012): 570-575.
}
