% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_gene_drug_response.R
\name{analyze_gene_drug_response_diff}
\alias{analyze_gene_drug_response_diff}
\title{Analyze Difference of Drug Response between Gene (Signature) High and Low Expression}
\usage{
analyze_gene_drug_response_diff(
  gene_list,
  drug = "ALL",
  tissue = "ALL",
  combine = FALSE,
  cutpoint = c(50, 50)
)
}
\arguments{
\item{gene_list}{a gene symbol list.}

\item{drug}{a drug name. Run example to check the list.}

\item{tissue}{a tissue name. Run example to check the list.}

\item{combine}{if \code{TRUE}, combine the expression of gene list as
a gene signature.}

\item{cutpoint}{cut point (in percent) for High and Low group, default is \code{c(50, 50)}.}
}
\value{
a \code{data.frame}.
}
\description{
Analyze Difference of Drug Response between Gene (Signature) High and Low Expression
}
\examples{
\dontrun{
analyze_gene_drug_response_diff("TP53")
analyze_gene_drug_response_diff(c("TP53", "KRAS"), drug = "AEW541")
analyze_gene_drug_response_diff(c("TP53", "KRAS"),
  tissue = "kidney",
  combine = TRUE
)

# Visualization
vis_gene_drug_response_diff("TP53")
}
}
