% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cross.R
\name{calc_pw_mut_cnv}
\alias{calc_pw_mut_cnv}
\title{Prepare Mut/CNV data of pathways given specific samples}
\usage{
calc_pw_mut_cnv(tumor_samples = NULL)
}
\arguments{
\item{tumor_samples}{Select specific tumor samples. Default NULL, indicating all tumor samples.}
}
\description{
Prepare Mut/CNV data of pathways given specific samples
}
