% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccle_info}
\alias{ccle_info}
\title{Phenotype info of CCLE database}
\format{
A \code{data.frame}
}
\source{
\href{https://xenabrowser.net/datapages/?dataset=ccle%2FCCLE_sample_info_file_2012-10-18.txt&host=https%3A%2F%2Fucscpublic.xenahubs.net&removeHub=https%3A%2F%2Fxena.treehouse.gi.ucsc.edu%3A443}{dataset link}.
}
\description{
Phenotype info of CCLE database
}
\examples{
data("ccle_info")
}
