% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccle_info_fine}
\alias{ccle_info_fine}
\title{Cleaned Phenotype Info of CCLE Database for grouping}
\format{
A \code{data.frame}
}
\source{
UCSC Xena.
}
\description{
Cleaned Phenotype Info of CCLE Database for grouping
}
\examples{
data("ccle_info_fine")
}
