% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cliMat}
\alias{cliMat}
\title{Toil Hub: TCGA survival data from paper}
\format{
A \link{data.frame}
}
\source{
Generate from paper
}
\description{
Toil Hub: TCGA survival data from paper
}
\examples{
data("tcga_clinicalMatrix")
}
