% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomSplitViolin.R
\name{create_quantile_segment_frame}
\alias{create_quantile_segment_frame}
\title{Visualize Violin plot}
\usage{
create_quantile_segment_frame(data, draw_quantiles, split = FALSE,
  grp = NULL)
}
\description{
Visualize Violin plot
}
