% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_datasets}
\alias{dat_datasets}
\title{Number of datasets in each cohort}
\format{A \link{data.frame}}
\source{
Generate from data-raw
}
\description{
Number of datasets in each cohort
}
\examples{
data("dat_datasets")
}
