% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_samples}
\alias{dat_samples}
\title{Number of samples in each cohort}
\format{A \link{data.frame}}
\source{
Generate from data-raw
}
\description{
Number of samples in each cohort
}
\examples{
data(dat_samples)
}
