% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_value.R
\docType{data}
\name{.opt_pancan}
\alias{.opt_pancan}
\title{A default setting for pan-cancer studies}
\format{
An object of class \code{list} of length 16.
}
\usage{
.opt_pancan
}
\description{
A default setting for pan-cancer studies
}
\keyword{datasets}
