% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomSplitViolin.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Visualize Violin plot}
\usage{
geom_split_violin(mapping = NULL, data = NULL, stat = "ydensity",
  position = "identity", ..., draw_quantiles = NULL, trim = TRUE,
  scale = "area", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\description{
Visualize Violin plot
}
