% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomSplitViolin.R
\docType{data}
\name{GeomSplitViolin}
\alias{GeomSplitViolin}
\title{Visualize Violin plot}
\format{An object of class \code{GeomSplitViolin} (inherits from \code{GeomViolin}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.}
\usage{
GeomSplitViolin
}
\description{
Visualize Violin plot
}
\keyword{datasets}
