% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nonomics_value.R
\name{get_nonomics_value}
\alias{get_nonomics_value}
\title{Fetch non-omics data of all samples from relevant databases}
\usage{
get_nonomics_value(
  database = c("toil", "pcawg", "ccle"),
  type = c("immune", "pathway", "tumor_index"),
  subtype = NULL
)
}
\arguments{
\item{database}{one of "toil" (tcga), "pcawg", "ccle"}

\item{type}{one of "immune", "pathway", "tumor_index"}

\item{subtype}{For "immune" type of toil/pcawg database, one of "CIBERSORT", "CIBERSORT-ABS", "EPIC", "MCPCOUNTER", "QUANTISEQ", "TIMER", "XCELL".
For "pathway" type of toil/pcawg database, one of "HALLMARK", "KEGG", "IOBR".
For "tumor_index" type of toil database, one of "Tumor_Purity","Tumor_Stemness","Tumor_Mutation_Burden","Microsatellite_Instability", "Genome_Instability".
For "tumor_index" type of pcawg/ccle database, set NULL.
If subtype is NULL, return all subtypes.}
}
\value{
dataframe
}
\description{
Fetch non-omics data of all samples from relevant databases
}
\examples{
\dontrun{
tcga_immune_xcell = get_nonomics_value(database = "toil", type = "immune", subtype = "XCELL")

tcga_pathway_kegg = get_nonomics_value(database = "toil", type = "pathway", subtype = "KEGG")

# Combined with sample information

head(tcga_clinical_fine)
head(tcga_surv)

head(pcawg_info_fine)

head(ccle_info_fine)

}

}
