% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_identifier.R
\name{get_signature_data}
\alias{get_signature_data}
\title{Get Molecule or Signature Data Values from Dense Matrix Dataset of UCSC Xena Data Hubs}
\usage{
get_signature_data(dataset, signature, host = NULL)
}
\arguments{
\item{dataset}{a UCSC Xena dataset in dense matrix format (rows are features
(e.g., gene, cell line) and columns are samples).}

\item{signature}{a molecular identifier (e.g., "TP53") or a formula specifying
genomic signature (\code{"TP53 + 2 * KRAS - 1.3 \\* PTEN"}).
\strong{NOTE}, when a signature is specified, a space must exist in the input.}

\item{host}{a UCSC Xena host, default is \code{NULL}, auto-detect from the dataset.}
}
\value{
a named vector.
}
\description{
Get Molecule or Signature Data Values from Dense Matrix Dataset of UCSC Xena Data Hubs
}
\examples{
\dontrun{
dataset <- "ccle/CCLE_copynumber_byGene_2013-12-03"
x <- get_signature_data(dataset, "TP53")
head(x)

signature <- "TP53 + 2*KRAS - 1.3*PTEN" # a space must exist in the string
y <- get_signature_data(dataset, signature)
head(y)
}
}
