% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immune_sig}
\alias{immune_sig}
\title{TCGA: immune signature data}
\format{
A \link{data.frame}
}
\source{
\url{https://gdc.cancer.gov/about-data/publications/panimmune}
}
\description{
TCGA: immune signature data
}
\examples{
data("immune_sig")
}
