% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_cat_cols}
\alias{keep_cat_cols}
\title{Keep Only Columns Used for Sample Selection}
\usage{
keep_cat_cols(x, keep_sam_cols = TRUE, return_idx = TRUE)
}
\arguments{
\item{x}{a \code{data.frame} with many columns.}

\item{keep_sam_cols}{if \code{TRUE} (default), keep columns with
pattern 'sample', 'patient', etc.}

\item{return_idx}{if \code{TRUE} (default), return index of 5 (at most) columns, it is useful in Shiny.}
}
\value{
a \code{data.frame} or a \code{list}.
}
\description{
Keep Only Columns Used for Sample Selection
}
