% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Dataset Provided by This Package}
\usage{
load_data(name)
}
\arguments{
\item{name}{dataset name, can be one of
TCGA.organ
ccle_absolute
ccle_info
pancan_MSI
tcga_MSI
tcga_TIL
tcga_chr_alteration
tcga_clinical
tcga_genome_instability
tcga_gtex
tcga_pan_immune_signature
tcga_purity
tcga_stemness
tcga_subtypes
tcga_surv
tcga_tmb
toil_info}
}
\value{
a dataset, typically a \code{data.frame}.
}
\description{
Load data from builtin or zenodo.
}
\examples{
load_data("tcga_surv")
}
