% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_general.R
\name{mol_quick_analysis}
\alias{mol_quick_analysis}
\title{Quick molecule analysis and report generation}
\usage{
mol_quick_analysis(molecule, data_type, out_dir = ".", out_report = FALSE)
}
\arguments{
\item{molecule}{a molecular identifier (e.g., "TP53") or a formula specifying
genomic signature (\code{"TP53 + 2 * KRAS - 1.3 * PTEN"}).}

\item{data_type}{data type. Can be one of "mRNA", "transcript", "protein",
"mutation", "cnv", "methylation", "miRNA".}

\item{out_dir}{path to save analysis result and report, default is '.'}

\item{out_report}{logical value wheather to generate html report}
}
\value{
a list.
}
\description{
Quick molecule analysis and report generation
}
