% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcawg_info}
\alias{pcawg_info}
\title{Phenotype Info of PCAWG Database}
\format{
A \code{data.frame}
}
\source{
UCSC Xena.
}
\description{
Phenotype Info of PCAWG Database
}
\examples{
data("pcawg_info")
}
