% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcawg_info_fine}
\alias{pcawg_info_fine}
\title{Cleaned Phenotype Info of PCAWG Database for grouping}
\format{
A \code{data.frame}
}
\source{
UCSC Xena.
}
\description{
Cleaned Phenotype Info of PCAWG Database for grouping
}
\examples{
data("pcawg_info_fine")
}
