% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purity_data}
\alias{purity_data}
\title{TCGA: purity data}
\format{
A \link{data.frame}
}
\source{
\url{https://www.nature.com/articles/ncomms9971#Sec14}
}
\description{
TCGA: purity data
}
\examples{
data("purity_data")
}
