% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_value.R
\name{query_pancan_value}
\alias{query_pancan_value}
\title{Query Single Identifier or Signature Value from Pan-cancer Database}
\usage{
query_pancan_value(
  molecule,
  data_type = c("mRNA", "transcript", "protein", "mutation", "cnv", "cnv_gistic2",
    "methylation", "miRNA"),
  database = c("toil", "ccle"),
  reset_id = NULL
)
}
\arguments{
\item{molecule}{a molecular identifier (e.g., "TP53") or a formula specifying
genomic signature (\code{"TP53 + 2 * KRAS - 1.3 * PTEN"}).}

\item{data_type}{data type. Can be one of "mRNA", "transcript", "protein",
"mutation", "cnv" (-2, -1, 0, 1, 2), "cnv_gistic2", "methylation", "miRNA".}

\item{database}{database, either 'toil' for TCGA TARGET GTEx, or 'ccle' for
CCLE.}
}
\value{
a list.
}
\description{
Query Single Identifier or Signature Value from Pan-cancer Database
}
