% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tcga_group.R
\name{query_tcga_group}
\alias{query_tcga_group}
\title{Group TCGA samples by build-in or custom phenotype and support filtering or merging operations}
\usage{
query_tcga_group(
  cancer = NULL,
  custom = NULL,
  group = "Gender",
  filter_by = NULL,
  merge_by = NULL,
  return_all = FALSE
)
}
\arguments{
\item{cancer}{select cancer cohort(s)}

\item{custom}{upload custom phenotype data}

\item{group}{target group names}

\item{filter_by}{filter samples by one or multiple criterion}

\item{merge_by}{merge the target group for main categories}

\item{return_all}{return the all phenotype data}
}
\value{
a list object with grouping samples and statistics
}
\description{
Group TCGA samples by build-in or custom phenotype and support filtering or merging operations
}
\examples{
\dontrun{
query_tcga_group(group = "Age")

query_tcga_group(cancer="BRCA", 
              group = "Stage_ajcc"
              )

query_tcga_group(cancer="BRCA", 
              group = "Stage_ajcc",
              filter_by = list(
                c("Code",c("TP"),"+"),
                c("Stage_ajcc",c(NA),"-"))
              )

query_tcga_group(cancer="BRCA", 
              group = "Stage_ajcc",
              merge_by = list(
                "Early"=c("Stage I"), 
                "Late" = c("Stage II","Stage III","Stage IV"))
              )

query_tcga_group(cancer="BRCA", 
              group = "Age",
              merge_by = list(
                "Young"= 60, 
                "Middle"= 80, 
                "Old" = 80
              )
              )
}
}
