% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ope_pancan_value.R, R/query_toil_value_df.R
\name{query_toil_value_df}
\alias{query_toil_value_df}
\title{Obtain ToilHub Info for Single Molecule}
\usage{
query_toil_value_df(identifier = "TP53")

query_toil_value_df(identifier = "TP53")
}
\arguments{
\item{identifier}{a length-1 character representing a gene symbol, ensembl gene id, or probe id.
Gene symbol is highly recommended.}
}
\value{
a \code{tibble}

a \code{tibble}
}
\description{
Obtain ToilHub Info for Single Molecule

Obtain ToilHub Info for Single Gene
}
\examples{
\dontrun{
t <- query_toil_value_df()
t
}
\dontrun{
t <- query_toil_value_df()
t
}
}
