% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_value.R
\name{query_value}
\alias{query_value}
\title{Query Single Identifier Value from Pan-cancer Database}
\usage{
query_value(
  identifier,
  data_type = c("mRNA", "transcript", "protein", "mutation", "cnv", "cnv_gistic2",
    "methylation", "miRNA"),
  database = c("toil", "ccle")
)
}
\arguments{
\item{identifier}{a ID, e.g. gene symbol.}

\item{data_type}{data type.}

\item{database}{database, either 'toil' for TCGA TARGET GTEx, or 'ccle' for
CCLE.}
}
\value{
a list.
}
\description{
Query Single Identifier Value from Pan-cancer Database
}
