% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stemness_data_RNA}
\alias{stemness_data_RNA}
\title{TCGA: Stemness data}
\format{
A \link{data.frame}
}
\source{
\url{https://pancanatlas.xenahubs.net/download/StemnessScores_DNAmeth_20170210.tsv.gz}
}
\description{
TCGA: Stemness data
}
\examples{
data("stemness_data_RNA")
}
