% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_clinical}
\alias{tcga_clinical}
\title{Toil Hub: TCGA Clinical Data}
\format{
A \link{data.frame}
}
\source{
Generate from data-raw
}
\description{
See \code{tcga_surv} for TCGA survival data.
}
\examples{
data("tcga_clinical")
}
