% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_clinical_fine}
\alias{tcga_clinical_fine}
\title{Toil Hub: Cleaned TCGA Clinical Data for grouping}
\format{
A \link{data.frame}
}
\source{
Generate from data-raw
}
\description{
See \code{tcga_surv} for TCGA survival data.
}
\examples{
data("tcga_clinical_fine")
}
