% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_genome_instability}
\alias{tcga_genome_instability}
\title{TCGA: genome instability data}
\format{
A \link{data.frame}
}
\source{
\url{https://gdc.cancer.gov/about-data/publications/PanCanStemness-2018}
}
\description{
TCGA: genome instability data
}
\examples{
data("tcga_genome_instability")
}
