% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_gtex}
\alias{tcga_gtex}
\title{Toil Hub: Merged TCGA GTEx Selected Phenotype}
\format{
A \link{data.frame}
}
\description{
Toil Hub: Merged TCGA GTEx Selected Phenotype
}
\examples{
data("tcga_gtex")
}
