% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_pan_tcga_pan_immune_signaturenature}
\alias{tcga_pan_tcga_pan_immune_signaturenature}
\title{TCGA: immune signature data}
\format{
A \link{data.frame}
}
\source{
\url{https://gdc.cancer.gov/about-data/publications/panimmune}
}
\description{
TCGA: immune signature data
}
\examples{
data("tcga_pan_tcga_pan_immune_signaturenature")
}
