% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_stemness}
\alias{tcga_stemness}
\title{TCGA: Stemness data}
\format{
A \link{data.frame}
}
\source{
\url{https://pancanatlas.xenahubs.net}
}
\description{
TCGA: Stemness data
}
\examples{
data("tcga_stemness")
}
