% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_surv}
\alias{tcga_surv}
\title{Toil Hub: TCGA Survival Data}
\format{
A \link{data.frame}
}
\source{
Generate from data-raw
}
\description{
Toil Hub: TCGA Survival Data
}
\examples{
data("tcga_surv")
}
