% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGAsubtype}
\alias{TCGAsubtype}
\title{TCGA subtype data}
\format{
A \link{data.frame}
}
\source{
\href{https://xenabrowser.net/datapages/?dataset=TCGASubtype.20170308.tsv&host=https%3A%2F%2Fpancanatlas.xenahubs.net&removeHub=https%3A%2F%2Fxena.treehouse.gi.ucsc.edu%3A443}{Molecular subtype}
and \href{https://xenabrowser.net/datapages/?dataset=Subtype_Immune_Model_Based.txt&host=https%3A%2F%2Fpancanatlas.xenahubs.net&removeHub=https%3A%2F%2Fxena.treehouse.gi.ucsc.edu%3A443}{immune subtype}.
}
\description{
TCGA subtype data
}
\examples{
data("TCGAsubtype")
}
