% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmb_data}
\alias{tmb_data}
\title{TCGA: TMB (tumor mutation burden) data}
\format{
A \link{data.frame}
}
\source{
\url{https://gdc.cancer.gov/about-data/publications/panimmune}
}
\description{
TCGA: TMB (tumor mutation burden) data
}
\examples{
data("tmb_data")
}
