% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toil_info}
\alias{toil_info}
\title{Toil Hub: TCGA TARGET GTEX selected phenotypes}
\format{A \link{data.frame}}
\source{
Generate from data-raw
}
\description{
Toil Hub: TCGA TARGET GTEX selected phenotypes
}
\examples{
data("toil_info")
}
