% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toil_surv}
\alias{toil_surv}
\title{Toil Hub: TCGA survival data}
\format{A \link{data.frame}}
\source{
Generate from data-raw
}
\description{
Toil Hub: TCGA survival data
}
\examples{
data("toil_surv")
}
