% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xenashiny.R
\docType{package}
\name{UCSCXenaShiny-package}
\alias{UCSCXenaShiny}
\alias{UCSCXenaShiny-package}
\title{Xena Shiny App}
\description{
A Shiny App for UCSC Xena Data Hubs. See \url{https://github.com/openbiox/UCSCXenaShiny} for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openbiox/UCSCXenaShiny}
  \item \url{https://openbiox.github.io/UCSCXenaShiny/}
  \item Report bugs at \url{https://github.com/openbiox/UCSCXenaShiny/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

Authors:
\itemize{
  \item Shensuo Li \email{lishensuo@163.com}
  \item Yi Xiong \email{neuroxiong@openbiox.com} (\href{https://orcid.org/0000-0002-4370-9824}{ORCID})
  \item Longfei Zhao \email{longfei8533@126.com} (\href{https://orcid.org/0000-0002-6277-0137}{ORCID})
  \item Kai Gu \email{gukai1212@163.com} (\href{https://orcid.org/0000-0002-0177-0774}{ORCID})
  \item Yin Li \email{yinli18@fudan.edu.cn}
  \item Fei Zhao \email{zhaofei415@mails.ucas.edu.cn}
}

}
\keyword{internal}
