% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cross.R
\name{vis_ccle_gene_cross_omics}
\alias{vis_ccle_gene_cross_omics}
\title{Visualize cross-omics of one gene in CCLE}
\usage{
vis_ccle_gene_cross_omics(
  gene = "TP53",
  tumor_projects = NULL,
  return_list = FALSE
)
}
\arguments{
\item{gene}{a gene symbol identifier (e.g., "TP53")}

\item{tumor_projects}{Select specific CCLE tissues. Default NULL, indicating all.}

\item{return_list}{TRUE returns a list including plot object and data. FALSE just returns plot.}
}
\value{
funkyheatmap
}
\description{
Visualize cross-omics of one gene in CCLE
}
