% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_ccle_value.R
\name{vis_ccle_tpm}
\alias{vis_ccle_tpm}
\title{Visualize CCLE Gene Expression in TPM Format}
\usage{
vis_ccle_tpm(Gene = "TP53", x.axis = "Type")
}
\arguments{
\item{Gene}{Gene symbal for comparision}

\item{x.axis}{Different parameters for x.axis}
}
\value{
a \code{ggplot} object
}
\description{
Visualize CCLE Gene Expression in TPM Format
}
