% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_dim_dist.R
\name{vis_dim_dist}
\alias{vis_dim_dist}
\title{Visualize the distribution difference of samples after dimension reduction analysis}
\usage{
vis_dim_dist(
  ids = c("TP53", "KRAS", "PTEN", "MDM2", "CDKN1A"),
  data_type = "mRNA",
  cancer_type = "LIHC",
  Mode = c("TumorNomral", "Clinical", "Cancer", "Custom"),
  group_name = NULL,
  custom_meta = NULL,
  TCGA.only = TRUE,
  return.data = FALSE,
  DR_method = c("PCA", "UMAP", "tSNE"),
  palette = "Set1",
  add_margin = NULL
)
}
\arguments{
\item{ids}{molecular identifiers (>=3)}

\item{data_type}{molecular types, refer to query_pancan_value() fuuntion}

\item{cancer_type}{select cancer cohort(s)}

\item{Mode}{how to define group metadata}

\item{group_name}{the provided phenotype under 'Clinical' mode. We collected six phenotypes based on TGCA project.
("age","gender","race","stage_ajcc","stage_clinical","histological_grade"). For age, the older and yunger groups are determined by the median value.
For other phenotypes (such as stage_ajcc), the primary phenotypes were substracted.}

\item{custom_meta}{the provided metadata under 'Custom' mode}

\item{TCGA.only}{whether to include GTEx normal samples under 'TumorNomral' mode}

\item{return.data}{whether to reture the raw meta/matrix data (list) instead of plot}

\item{DR_method}{the dimension reduction method}

\item{palette}{the color setting of RColorBrewer}

\item{add_margin}{the marginal plot (NULL, "density", "boxplot")}
}
\value{
a \code{ggplot} object.
}
\description{
Visualize the distribution difference of samples after dimension reduction analysis
}
\examples{
\donttest{
# The examples work, but may take a long time to run.

p <- vis_dim_dist(DR_method = "UMAP")
p <- vis_dim_dist(Mode = "Clinical", group_name = "age")

pancan_identifiers <- readRDS(
  system.file(
    "extdata", "pancan_identifier_list.rds",
    package = "UCSCXenaShiny"
  )
)
set.seed(42)
ids <- sample(pancan_identifiers$miRNA, 20)
p <- vis_dim_dist(ids = ids, data_type = "miRNA")

p <- vis_dim_dist(cancer_type = c("LUAD", "LIHC", "LUSC"), Mode = "Cancer")

set.seed(42)
custom_meta <- tcga_gtex \%>\%
  dplyr::mutate(across(where(is.factor), as.character)) \%>\%
  dplyr::filter(tissue == "LUAD") \%>\%
  dplyr::filter(grepl("TCGA", sample)) \%>\%
  dplyr::sample_frac(0.8) \%>\%
  dplyr::mutate(group = sample(c("TA", "TB"), size = nrow(.), replace = TRUE)) \%>\%
  dplyr::select(sample, group)
p <- vis_dim_dist(cancer_type = c("LUAD"), Mode = "Custom", custom_meta = custom_meta)
}
}
