% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_analyze_gene_drug_response.R
\name{vis_gene_drug_response_asso}
\alias{vis_gene_drug_response_asso}
\title{Visualize Gene and Drug-Target Association with CCLE Data}
\usage{
vis_gene_drug_response_asso(
  Gene = "TP53",
  x_axis_type = c("mean.diff", "median.diff"),
  output_form = c("plotly", "ggplot2")
)
}
\arguments{
\item{Gene}{a molecular identifier (e.g., "TP53") or a formula specifying
genomic signature (\code{"TP53 + 2 * KRAS - 1.3 * PTEN"}).}

\item{x_axis_type}{set the value type for X axis.}

\item{output_form}{\code{plotly} or \code{ggplot2}.}
}
\value{
\code{plotly} or \code{ggplot2} object.
}
\description{
See \link{analyze_gene_drug_response_asso} for examples.
}
