% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_gene_drug_response2.R
\name{vis_gene_drug_response_asso2}
\alias{vis_gene_drug_response_asso2}
\title{Another Function to Visualize Gene and Drug-Target Association with CCLE Data}
\usage{
vis_gene_drug_response_asso2(
  Gene = "TP53",
  response_type = c("IC50", "EC50", "Amax", "ActArea"),
  tissue = "ALL",
  manova = FALSE,
  output_form = c("plotly", "ggplot2")
)
}
\arguments{
\item{Gene}{a gene symbol (list), when it is a list, all
genes are combined as a signature.}

\item{response_type}{one measure type for drug response.}

\item{tissue}{a tissue name. Check examples.}

\item{output_form}{\code{plotly} or \code{ggplot2}.}
}
\value{
\code{plotly} or \code{ggplot2} object.
}
\description{
See \link{analyze_gene_drug_response_asso2} for examples.
}
