% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_analyze_gene_drug_response.R
\name{vis_gene_drug_response_diff}
\alias{vis_gene_drug_response_diff}
\title{Visualize Gene and Drug Response Difference}
\usage{
vis_gene_drug_response_diff(
  Gene = "TP53",
  tissue = "prostate",
  Show.P.value = TRUE,
  Show.P.label = TRUE,
  Method = "wilcox.test"
)
}
\arguments{
\item{Gene}{a molecular identifier (e.g., "TP53") or a formula specifying
genomic signature (\code{"TP53 + 2 * KRAS - 1.3 * PTEN"}).}

\item{tissue}{select cell line origin tissue.}

\item{Show.P.value}{\code{TRUE} or \code{FALSE} whether to count P value}

\item{Show.P.label}{\code{TRUE} or \code{FALSE} present p value with number or label \code{*}, \verb{**}, \verb{***} and \verb{****}}

\item{Method}{default method is wilcox.test}
}
\value{
a \code{ggplot} object.
}
\description{
See \link{analyze_gene_drug_response_diff} for examples.
}
