% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pancan_value.R
\name{vis_gene_immune_cor}
\alias{vis_gene_immune_cor}
\title{Heatmap for Correlation between Gene and Immune Signatures}
\usage{
vis_gene_immune_cor(
  Gene = "TP53",
  Cor_method = "spearman",
  data_type = "mRNA",
  Immune_sig_type = "Cibersort"
)
}
\arguments{
\item{Gene}{Gene symbal for comparision}

\item{Cor_method}{correlation method}

\item{data_type}{data type. Can be one of "mRNA", "transcript", "protein",
"mutation", "cnv" (-2, -1, 0, 1, 2), "cnv_gistic2", "methylation", "miRNA".}

\item{Immune_sig_type}{quantification method, default is "Cibersort"}
}
\description{
Heatmap for Correlation between Gene and Immune Signatures
}
\examples{
\donttest{
p <- vis_gene_immune_cor(Gene = "TP53")
}
}
