% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_gene_pw_cor.R
\name{vis_gene_pw_cor}
\alias{vis_gene_pw_cor}
\title{Visualize Correlation between Gene and Pathway signature Score}
\usage{
vis_gene_pw_cor(
  Gene = "TP53",
  data_type = "mRNA",
  pw_name = NULL,
  Cancer = "ACC",
  cor_method = "spearman",
  cor_cutoff = list(r = 0.3, p = 0.01),
  plot = TRUE
)
}
\arguments{
\item{Gene}{a molecular identifier (e.g., "TP53") or a formula
specifying genomic signature ("TP53 + 2 * KRAS - 1.3 * PTEN").}

\item{data_type}{choose gene profile type, including
"mRNA", "transcript", "protein", "mutation", "cnv" (-2, -1, 0, 1, 2), "cnv_gistic2", "methylation", "miRNA".}

\item{pw_name}{the queried Pathway name, see the supported pathway from 'load("toil_sig_score")'default is NULL}

\item{Cancer}{select cancer cohort(s)}

\item{cor_method}{select correlation coefficient (pearson/spearman)}

\item{cor_cutoff}{a list with 2 elements names 'r' and 'p' to define significant correlations}

\item{plot}{output the plot directly (default) or raw data.frame}
}
\value{
a \code{ggplot} object or  dataframe
}
\description{
Visualize Correlation between Gene and Pathway signature Score
}
\examples{
p1_1 = vis_gene_pw_cor(Gene="TP53", data_type="mRNA", Cancer='ACC')
p1_1_data = vis_gene_pw_cor(Gene="CD4", data_type="mRNA", Cancer='ACC', plot=FALSE)
p1_2 = vis_gene_pw_cor(Gene="hsa-miR-1228-3p", data_type="miRNA", Cancer='ACC')


p2 = vis_gene_pw_cor(Gene="TP53", data_type="mRNA", Cancer='ACC', 
                     pw_name="KEGG_VEGF_SIGNALING_PATHWAY")

p3 = vis_gene_pw_cor(Gene="TP53", data_type="mRNA", Cancer='Overall', 
                     pw_name="KEGG_VEGF_SIGNALING_PATHWAY")
}
