% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pancan_value.R
\name{vis_gene_stemness_cor}
\alias{vis_gene_stemness_cor}
\title{Visualize Correlation between Gene and Tumor Stemness}
\usage{
vis_gene_stemness_cor(
  Gene = "TP53",
  Cor_method = "spearman",
  data_type = "mRNA"
)
}
\arguments{
\item{Gene}{input gene}

\item{Cor_method}{correlation method}

\item{data_type}{choose gene profile type, including "mRNA","transcript","methylation","miRNA","protein","cnv_gistic2"}
}
\description{
Visualize Correlation between Gene and Tumor Stemness
}
\examples{
\donttest{
p <- vis_gene_stemness_cor(Gene = "TP53")
}
}
