% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pancan_value.R
\name{vis_gene_TIL_cor}
\alias{vis_gene_TIL_cor}
\title{Heatmap for correlation between gene and tumor immune signatures}
\usage{
vis_gene_TIL_cor(
  Gene = "TP53",
  Cor_method = "spearman",
  sig = c("B cell_TIMER", "T cell CD4+_TIMER", "T cell CD8+_TIMER", "Neutrophil_TIMER",
    "Macrophage_TIMER", "Myeloid dendritic cell_TIMER")
)
}
\arguments{
\item{Gene}{Gene symbal for comparision}

\item{Cor_method}{correlation method}

\item{sig}{Immune Signature, default: result from TIMER}
}
\description{
Heatmap for correlation between gene and tumor immune signatures
}
\examples{
\donttest{
p <- vis_gene_TIL_cor(Gene = "TP53")
}
}
