% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_identifier.R
\name{vis_identifier_cor}
\alias{vis_identifier_cor}
\title{Visualize Identifier-identifier Correlation}
\usage{
vis_identifier_cor(id1, id2, dataset1, dataset2, samples = NULL, ...)
}
\arguments{
\item{id1}{the first molecule identifier.}

\item{id2}{the second molecule identifier.}

\item{dataset1}{the dataset to obtain \code{id1}.}

\item{dataset2}{the dataset to obtain \code{id2}.}
}
\description{
Visualize Identifier-identifier Correlation
}
