% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pancan_value.R
\name{vis_pancan_anatomy}
\alias{vis_pancan_anatomy}
\title{Visualize single gene expression in anatomy location}
\usage{
vis_pancan_anatomy(Gene = "TP53", Gender = c("Female", "Male"))
}
\arguments{
\item{Gene}{Gene symbal for comparision}

\item{Gender}{a string, "Female" (default) or "Male".}
}
\value{
a \code{ggplot} object
}
\description{
Visualize single gene expression in anatomy location
}
