% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cross.R
\name{vis_pcawg_gene_cross_omics}
\alias{vis_pcawg_gene_cross_omics}
\title{Visualize cross-omics of one gene in PCAWG}
\usage{
vis_pcawg_gene_cross_omics(
  gene = "TP53",
  tumor_projects = NULL,
  n_promoter = 0,
  add_mean_promoter = FALSE,
  promoter_type = c("relative", "raw", "outlier"),
  return_list = FALSE
)
}
\arguments{
\item{gene}{a gene symbol identifier (e.g., "TP53")}

\item{tumor_projects}{Select specific PCAWG projects. Default NULL, indicating all.}

\item{n_promoter}{specific promoter identifiers or number of sampling.}

\item{add_mean_promoter}{whether to add median promoter activity.}

\item{promoter_type}{one of "relative", "raw", "outlier".}

\item{return_list}{TRUE returns a list including plot object and data. FALSE just returns plot.}
}
\value{
funkyheatmap
}
\description{
Visualize cross-omics of one gene in PCAWG
}
