% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cross.R
\name{vis_pcawg_pathway_cross_omics}
\alias{vis_pcawg_pathway_cross_omics}
\title{Visualize cross-omics of one pathway in PCAWG}
\usage{
vis_pcawg_pathway_cross_omics(
  pw = "HALLMARK_ADIPOGENESIS",
  tumor_projects = NULL,
  return_list = FALSE
)
}
\arguments{
\item{pw}{pathway name}

\item{tumor_projects}{Select specific PCAWG projects. Default NULL, indicating all.}

\item{return_list}{TRUE returns a list including plot object and data. FALSE just returns plot.}
}
\value{
funkyheatmap
}
\description{
Visualize cross-omics of one pathway in PCAWG
}
