% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{app_run}
\alias{app_run}
\title{Run UCSC Xena Shiny App}
\usage{
app_run(runMode = "client")
}
\arguments{
\item{runMode}{default is 'client' for personal user, set it to 'server' for running on server.}
}
\description{
Run UCSC Xena Shiny App
}
\examples{
\dontrun{
app_run()
}
}
