% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_func.R
\name{batch_download}
\alias{batch_download}
\title{download data for shiny batch analysis}
\usage{
batch_download(
  L1,
  L2,
  L3,
  tumor_index_list,
  tcga_TIL,
  tcga_PW,
  opt_pancan,
  custom_metadata = NULL
)
}
\arguments{
\item{L1}{level 1  main datatype}

\item{L2}{level 2  sub datatype}

\item{L3}{level 3  identifier}

\item{tumor_index_list}{Tumor index data. See shiny App.R file}

\item{tcga_TIL}{Tumor immune infiltration  data. See shiny App.R file}

\item{tcga_PW}{Pathway activity data. See shiny App.R file}

\item{opt_pancan}{molecular datasets parameters}

\item{custom_metadata}{user customized metadata}
}
\description{
download data for shiny batch analysis
}
