% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccle_absolute}
\alias{ccle_absolute}
\title{ABSOLUTE Result of CCLE Database}
\format{
A \code{data.frame}
}
\source{
see "data_source" attribute.
}
\description{
ABSOLUTE Result of CCLE Database
}
\examples{
data("ccle_absolute")
}
