% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccle_info}
\alias{ccle_info}
\title{Phenotype info of CCLE database}
\format{
A \code{data.frame}
}
\source{
Generate from data-raw
}
\description{
Phenotype info of CCLE database
}
\examples{
data("ccle_info")
}
