% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccle_info}
\alias{ccle_info}
\title{Phenotype Info of CCLE Database}
\format{
A \code{data.frame}
}
\source{
UCSC Xena.
}
\description{
Phenotype Info of CCLE Database
}
\examples{
data("ccle_info")
}
