% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_gtex}
\alias{tcga_gtex}
\title{Toil Hub: merged TCGA GTEX selected phenotypes}
\format{A \link{data.frame}}
\description{
Toil Hub: merged TCGA GTEX selected phenotypes
}
\examples{
data("tcga_gtex_sampleinfo")
}
