% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_tmb}
\alias{tcga_tmb}
\title{TCGA: TMB (Tumor Mutation Burden) Data}
\format{
A \link{data.frame}
}
\source{
\url{https://gdc.cancer.gov/about-data/publications/panimmune}
}
\description{
TCGA: TMB (Tumor Mutation Burden) Data
}
\examples{
data("tcga_tmb")
}
