% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_ccle_value.R
\name{vis_ccle_gene_cor}
\alias{vis_ccle_gene_cor}
\title{Visualize CCLE Gene Expression Correlation}
\usage{
vis_ccle_gene_cor(
  Gene1 = "CSF1R",
  Gene2 = "JAK3",
  data_type1 = "mRNA",
  data_type2 = "mRNA",
  cor_method = "spearman",
  use_log_x = FALSE,
  use_log_y = FALSE,
  use_regline = TRUE
)
}
\arguments{
\item{Gene1}{the first gene}

\item{Gene2}{the second gene}

\item{data_type1}{choose gene profile type for the first gene, including "mRNA","transcript","methylation","miRNA","protein","cnv_gistic2"}

\item{data_type2}{choose gene profile type for the second gene, including "mRNA","transcript","methylation","miRNA","protein","cnv_gistic2"}

\item{use_regline}{if \code{TRUE}, add regression line.}
}
\value{
a \code{ggplot} object
}
\description{
Visualize CCLE Gene Expression Correlation
}
