% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_dim_dist.R
\name{vis_dim_dist}
\alias{vis_dim_dist}
\title{Visualize the distribution difference of samples after dimension reduction analysis}
\usage{
vis_dim_dist(
  ids = c("TP53", "KRAS", "PTEN", "MDM2", "CDKN1A"),
  data_type = "mRNA",
  return.data = FALSE,
  DR_method = c("PCA", "UMAP", "tSNE"),
  palette = "Set1",
  add_margin = NULL,
  group_levels = NULL,
  opt_pancan = .opt_pancan,
  ...
)
}
\arguments{
\item{ids}{molecular identifiers (>=3)}

\item{data_type}{molecular types, refer to query_pancan_value() fuuntion}

\item{return.data}{whether to reture the raw meta/matrix data (list) instead of plot}

\item{DR_method}{the dimension reduction method}

\item{palette}{the color setting of RColorBrewer}

\item{add_margin}{the marginal plot (NULL, "density", "boxplot")}

\item{group_levels}{group levels setting}

\item{opt_pancan}{specify one dataset for some molercular profiles}

\item{...}{parameters refer to query_tcga_group() function}
}
\value{
a ggplot object or rawdata list
}
\description{
Visualize the distribution difference of samples after dimension reduction analysis
}
\examples{
\dontrun{
vis_dim_dist(
  ids = c("TP53", "KRAS", "PTEN", "MDM2", "CDKN1A"),
  cancer = "BRCA",
  group = "Gender",
  group_levels = NULL
)


vis_dim_dist(
  ids = c("TP53", "KRAS", "PTEN", "MDM2", "CDKN1A"),
  cancer = "BRCA",
  group = "Code",
  merge_by = list("Tumor"=c("TP","TM"),"Normal"=c("NT")),
  group_levels = c("Normal","Tumor"),
  add_margin = "boxplot"
)

vis_dim_dist(
  ids = c("TP53", "KRAS", "PTEN", "MDM2", "CDKN1A"),
  cancer = c("LIHC","BRCA"),
  group = "Cancer",
  group_levels = c("BRCA","LIHC"),
  filter_by = list(c("Code",c("TP"),"+")),
  add_margin = "density"
)
}

}
