% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pancan_value.R
\name{vis_gene_cor_cancer}
\alias{vis_gene_cor_cancer}
\title{Visualize Gene-Gene Correlation in a TCGA Cancer Type}
\usage{
vis_gene_cor_cancer(
  Gene1 = "CSF1R",
  Gene2 = "JAK3",
  purity_adj = TRUE,
  split = FALSE,
  cancer_choose = "GBM",
  cor_method = "spearman"
)
}
\arguments{
\item{Gene1}{the first gene.}

\item{Gene2}{the second gene.}

\item{purity_adj}{whether performing partial correlation adjusted by purity.}

\item{split}{whether split by TCGA tumor tissue.}

\item{cancer_choose}{TCGA cohort name, e.g. "ACC".}

\item{cor_method}{correlation method.}
}
\description{
Visualize Gene-Gene Correlation in a TCGA Cancer Type
}
