% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pancan_value.R
\name{vis_gene_TIL_cor}
\alias{vis_gene_TIL_cor}
\title{Heatmap for Correlation between Gene and Tumor Immune Infiltration (TIL)}
\usage{
vis_gene_TIL_cor(
  Gene = "TP53",
  Cor_method = "spearman",
  data_type = "mRNA",
  sig = c("B cell_TIMER", "T cell CD4+_TIMER", "T cell CD8+_TIMER", "Neutrophil_TIMER",
    "Macrophage_TIMER", "Myeloid dendritic cell_TIMER")
)
}
\arguments{
\item{Gene}{Molecule identifier (gene symbol in most cases)}

\item{Cor_method}{correlation method}

\item{data_type}{data type. Can be one of "mRNA", "transcript", "protein",
"mutation", "cnv" (-2, -1, 0, 1, 2), "cnv_gistic2", "methylation", "miRNA".}

\item{sig}{Immune Signature, default: result from TIMER}
}
\description{
Heatmap for Correlation between Gene and Tumor Immune Infiltration (TIL)
}
\examples{
\donttest{
p <- vis_gene_TIL_cor(Gene = "TP53")
}
}
