% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pancan_value.R
\name{vis_gene_tmb_cor}
\alias{vis_gene_tmb_cor}
\title{Visualize TMB and gene}
\usage{
vis_gene_tmb_cor(Gene = "TP53", Cor_method = "spearman")
}
\arguments{
\item{Gene}{Gene symbal for comparision}

\item{Cor_method}{correlation method}
}
\description{
Visualize TMB and gene
}
\examples{
\donttest{
p <- vis_gene_tmb_cor(Gene = "TP53")
}
}
