% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pancan_value.R
\name{vis_unicox_tree}
\alias{vis_unicox_tree}
\title{Visualize single gene uni-cox result}
\usage{
vis_unicox_tree(Gene = "TP53")
}
\arguments{
\item{Gene}{Gene symbal for comparision}
}
\value{
a \code{ggplot} object
}
\description{
Visualize single gene uni-cox result
}
\examples{
\donttest{
p <- vis_unicox_tree(Gene = "TP53")
}
}
