% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pancan_value.R
\name{vis_unicox_tree}
\alias{vis_unicox_tree}
\title{Visualize Single Gene Univariable Cox Result from Toil Data Hub}
\usage{
vis_unicox_tree(
  Gene = "TP53",
  measure = "OS",
  data_type = "mRNA",
  threshold = 0.5,
  values = c("grey", "#E31A1C", "#377DB8")
)
}
\arguments{
\item{Gene}{Molecule identifier (gene symbol in most cases)}

\item{measure}{a survival measure, e.g. "OS".}

\item{data_type}{data type. Can be one of "mRNA", "transcript", "protein",
"mutation", "cnv" (-2, -1, 0, 1, 2), "cnv_gistic2", "methylation", "miRNA".}

\item{threshold}{a expression cutoff, \code{0.5} for median.}

\item{values}{the color to fill tumor or normal}
}
\value{
a \code{ggplot} object
}
\description{
Visualize Single Gene Univariable Cox Result from Toil Data Hub
}
\examples{
\donttest{
p <- vis_unicox_tree(Gene = "TP53")
}
}
